@@ -7,57 +7,58 @@ environment:
   sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
-  battery_plus: ^5.0.3
+  battery_plus: ^6.2.1
   cross_file: ^0.3.3
   crypto: ^3.0.3
   collection: ^1.17.1
   dio: ^5.2.1
-  drift: ^2.18.0
+  drift: ^2.24.0
+  drift_flutter: ^0.2.4
   easy_refresh: ^3.3.2
   equatable: ^2.0.5
   expandable_sliver_list: ^4.1.1
-  file: ^6.0.0
-  flex_color_scheme: ^7.3.1
+  file: ^7.0.1
+  flex_color_scheme: ^8.1.0
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.2
-  flutter_cache_manager: ^3.3.1
+  flutter_bloc: ^9.0.0
+  flutter_cache_manager: ^3.4.1
   flutter_localizations:
     sdk: flutter
-  flutter_typeahead: ^4.7.0
+  flutter_typeahead: ^5.2.0
   geolocator: ^13.0.2
   holding_gesture: ^1.2.0
-  http: ^0.13.6
+  http: ^1.3.0
   intl: ^0.19.0
   # Используется в дополнение с geolocator из-за https://github.com/Baseflow/flutter-geolocator/issues/1166
-  location: ^5.0.3
+  location: ^8.0.0
   mask_text_input_formatter: ^2.4.0
-  package_info_plus: ^4.0.2
+  package_info_plus: ^8.1.3
   path: ^1.9.0
   path_provider: ^2.0.9
-  photo_view: ^0.14.0
+  photo_view: ^0.15.0
   provider: ^6.0.5
   pub_semver: ^2.1.1
   queue: ^3.1.0+2
-  quiver: ^3.0.1
+  quiver: ^3.2.1
   retry: ^3.1.2
-  rxdart: ^0.27.7
+  rxdart: ^0.28.0
   scroll_to_index: ^3.0.1
-  sentry_flutter: ^7.14.0
+  sentry_flutter: ^8.12.0
   sqlite3_flutter_libs: ^0.5.0
   u_app_utils:
     git:
       url: git@github.com:Unact/u_app_utils.git
-      ref: 5cc092d
+      ref: ffe4dd2
   url_launcher: ^6.3.1
   uuid: ^4.2.2
   yandex_mapkit: ^4.0.2
 
 dev_dependencies:
-  build_runner: ^2.4.9
-  drift_dev: ^2.18.0
-  flutter_launcher_icons: ^0.13.1
-  flutter_lints: ^2.0.1
+  build_runner: ^2.4.14
+  drift_dev: ^2.24.0
+  flutter_launcher_icons: ^0.14.3
+  flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
 
