@@ -130,7 +130,7 @@ part of 'database.dart';
   }
 )
 class OrdersDao extends DatabaseAccessor<AppDataStore> with _$OrdersDaoMixin {
-  OrdersDao(AppDataStore db) : super(db);
+  OrdersDao(super.db);
 
   Future<void> regenerateOrdersGuid() async {
     await db._regenerateGuid(orders);
@@ -239,7 +239,8 @@ class OrdersDao extends DatabaseAccessor<AppDataStore> with _$OrdersDaoMixin {
     required List<int>? goodsIds,
     required bool onlyLatest,
     required bool onlyForPhysical,
-    required bool onlyWithoutDocs
+    required bool onlyWithoutDocs,
+    required String? barcode
   }) async {
     final hasBonusProgram = existsQuery(
       select(goodsBonusPrograms)
@@ -253,6 +254,7 @@ class OrdersDao extends DatabaseAccessor<AppDataStore> with _$OrdersDaoMixin {
     final query = select(allGoods)
       ..where((tbl) => tbl.name.containsCase(name ?? ''))
       ..where((tbl) => tbl.extraLabel.containsCase(extraLabel ?? ''))
+      ..where((tbl) => tbl.barcodes.containsCase(barcode ?? ''))
       ..where(categoryId != null ? (tbl) => tbl.categoryId.equals(categoryId) : (tbl) => const Constant(true))
       ..where(bonusProgramId != null ? (tbl) => hasBonusProgram : (tbl) => const Constant(true))
       ..where((tbl) => hasStock)
@@ -387,7 +389,7 @@ class GoodsDetail {
       [
         (totalDiscount != null ? (1 - totalDiscount / 100) * pricelistPrice : null),
         (minCoef != null ? minCoef * goods.cost : null)
-      ].whereNotNull().maxOrNull ??
+      ].nonNulls.maxOrNull ??
       pricelistPrice;
 
     return bonusPrice.roundDigits(2);
