@@ -16,8 +16,8 @@ part 'home_view_model.dart';
 
 class HomePage extends StatelessWidget {
   HomePage({
-    Key? key
-  }) : super(key: key);
+    super.key
+  });
 
   @override
   Widget build(BuildContext context) {
