@@ -28,8 +28,8 @@ class OrderPage extends StatelessWidget {
 
   OrderPage({
     required this.orderEx,
-    Key? key
-  }) : super(key: key);
+    super.key
+  });
 
   @override
   Widget build(BuildContext context) {
