@@ -13,7 +13,6 @@ import 'package:u_app_utils/u_app_utils.dart';
 import '/app/constants/strings.dart';
 import '/app/data/database.dart';
 import '/app/entities/entities.dart';
-import '/app/extensions/io_file_system.dart';
 import '/app/repositories/base_repository.dart';
 import '/app/services/palman_api.dart';
 
@@ -31,7 +30,7 @@ class PointsRepository extends BaseRepository {
     ),
   );
 
-  PointsRepository(AppDataStore dataStore, RenewApi api) : super(dataStore, api);
+  PointsRepository(super.dataStore, super.api);
 
   Stream<List<PointBuyerRoutePoint>> watchPointBuyerRoutePoints() {
     return dataStore.pointsDao.watchPointBuyerRoutePoints();
