@@ -10,7 +10,6 @@ import 'package:u_app_utils/u_app_utils.dart';
 import '/app/constants/strings.dart';
 import '/app/data/database.dart';
 import '/app/entities/entities.dart';
-import '/app/extensions/io_file_system.dart';
 import '/app/repositories/base_repository.dart';
 import '/app/services/palman_api.dart';
 
@@ -40,7 +39,7 @@ class VisitsRepository extends BaseRepository {
     ),
   );
 
-  VisitsRepository(AppDataStore dataStore, RenewApi api) : super(dataStore, api);
+  VisitsRepository(super.dataStore, super.api);
 
   Stream<List<VisitSoftware>> watchVisitSoftwares(String visitGuid) {
     return dataStore.visitsDao.watchVisitSoftwares(visitGuid);
