@@ -6,11 +6,11 @@ class BuyerField extends StatefulWidget {
   final TextEditingController? controller;
 
   BuyerField({
-    Key? key,
+    super.key,
     required this.buyerExList,
     this.controller,
     required this.onSelect
-  }) : super(key: key);
+  });
 
   @override
   State<BuyerField> createState() => _BuyerFieldState();
@@ -18,39 +18,45 @@ class BuyerField extends StatefulWidget {
 
 class _BuyerFieldState extends State<BuyerField> {
   late final TextEditingController controller = widget.controller ?? TextEditingController();
+  final FocusNode focusNode = FocusNode();
 
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
 
     return TypeAheadField(
-      textFieldConfiguration: TextFieldConfiguration(
-        maxLines: 1,
-        cursorColor: theme.textSelectionTheme.selectionColor,
-        autocorrect: false,
-        controller: controller,
-        style: Styles.formStyle,
-        textInputAction: TextInputAction.search,
-        decoration: InputDecoration(
-          labelText: 'Клиент',
-          suffixIcon: controller.text == '' ? null : IconButton(
-            tooltip: 'Очистить',
-            onPressed: () {
-              controller.text = '';
-              widget.onSelect(null);
-            },
-            icon: const Icon(Icons.clear)
-          )
-        ),
-        onChanged: (value) => value.isEmpty ? widget.onSelect(null) : null
-      ),
+      focusNode: focusNode,
+      controller: controller,
+      builder: (context, controller, focusNode) {
+        return TextField(
+          focusNode: focusNode,
+          maxLines: 1,
+          cursorColor: theme.textSelectionTheme.selectionColor,
+          autocorrect: false,
+          controller: controller,
+          style: Styles.formStyle,
+          textInputAction: TextInputAction.search,
+          decoration: InputDecoration(
+            labelText: 'Клиент',
+            suffixIcon: controller.text == '' ? null : IconButton(
+              tooltip: 'Очистить',
+              onPressed: () {
+                controller.text = '';
+                widget.onSelect(null);
+              },
+              icon: const Icon(Icons.clear)
+            )
+          ),
+          onChanged: (value) => value.isEmpty ? widget.onSelect(null) : null
+        );
+      },
       errorBuilder: (BuildContext ctx, error) {
         return Padding(
           padding: const EdgeInsets.all(8),
           child: Text('Произошла ошибка', style: Styles.formStyle.copyWith(color: theme.colorScheme.error))
         );
       },
-      noItemsFoundBuilder: (BuildContext ctx) {
+      emptyBuilder: (BuildContext ctx) {
         return Padding(
           padding: const EdgeInsets.all(8),
           child: Text(
@@ -72,8 +78,9 @@ class _BuyerFieldState extends State<BuyerField> {
           subtitle: Text("${suggestion.site.name}, ${suggestion.partner.name}", style: Styles.tileText),
         );
       },
-      onSuggestionSelected: (BuyerEx suggestion) async {
+      onSelected: (BuyerEx suggestion) async {
         controller.text = suggestion.buyer.fullname;
+        focusNode.unfocus();
         widget.onSelect(suggestion.buyer);
       }
     );
